requirements.txt
langchain
sentence-transformers
rich
torch
pandas
transformers
tqdm
jsonschema
requests
numpy
python-dotenv
